убрал memo у otherproduct, добавил компонент product
